chore: update for LESS v3
